KASan: report: print shadow memory state to stderr
The hexdump is currently printed to stdout via printk in
print_hex_dump. KASAN reports are special: They should be printed to
stderr, but they should not be logged as that would involve
reallocation. Therefore instead of calling print_hex_dump, call eprintf
with the %*ph format specifier instead. This allows us to consolidate
the code some more.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240731071416.2558182-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>